+++ validation/jsoncodeblock.py
@@ -6,6 +6,8 @@ from urllib.parse import urljoin
 import pytest
 import jsonschema
 
+from validation import errors
+
 CurrentValidator = jsonschema.Draft7Validator
 
 # set up hardcoded global variable that will store all the jsonschema
@@ -18,11 +20,17 @@ spec_folder = 'specs'
 def get_all_json(path):
     for fname in Path(path).glob('**/*.json'):
         with open(fname) as json_file:
-            yield json.load(json_file)
+            try:
+                yield fname, json.load(json_file)
+            except json.decoder.JSONDecodeError as exc:
+                raise errors.InvalidSchemaError(fname) from exc
 
 def build_url_store():
-    for contents in get_all_json(Path(spec_folder)):
-        yield contents['$id'], contents
+    for fname, contents in get_all_json(Path(spec_folder)):
+        try:
+            yield contents['$id'], contents
+        except KeyError:
+            raise errors.MissingSchemaIdError(fname)
 
 prepopulated_cache = dict(build_url_store())
 
@@ -37,7 +45,7 @@ class JsonCodeblock(pytest.Item):
         try:
             extra_keys = json.loads(self.extra_info or '{}')
         except json.decoder.JSONDecodeError as exc:
-            raise InvalidInfoStringError(self.extra_info) from exc
+            raise errors.InvalidInfoStringError(self.extra_info) from exc
         body.update(extra_keys)
         self._validate(body)
 
@@ -59,43 +67,11 @@ class JsonCodeblock(pytest.Item):
         try:
             schema_uri = body['$schema']
         except KeyError as exc:
-            raise NoSchemaSpecifiedError() from exc
+            raise errors.NoSchemaSpecifiedError() from exc
         try:
             _, schema = resolver.resolve(schema_uri)
         except jsonschema.RefResolutionError as exc:
-            raise SchemaResolutionError(schema_uri) from exc
+            raise errors.SchemaResolutionError(schema_uri) from exc
         CurrentValidator.check_schema(schema)
         validator = CurrentValidator(schema, resolver=resolver)
         validator.validate(body)
-
-class JsonCodeblockError(Exception):
-    pass
-
-class SchemaResolutionError(JsonCodeblockError):
-    def __init__(self, schema):
-        self.schema = schema
-
-    def __str__(self):
-        return self.schema
-
-class NoSchemaSpecifiedError(JsonCodeblockError):
-    def __str__(self):
-        return (
-            'Any codeblock that is labelled as "json" must specify a'
-            ' a schema. Schemas can be specified with the "$schema"'
-            ' key in the body of the codeblock or as a key in an'
-            ' object after "json" in the info string (which will be'
-            ' merged with the object in the body), like so:\n'
-            '``` json {"$schema": "https://example.com/schema"}\n'
-            '{}\n```'
-        )
-
-class InvalidInfoStringError(JsonCodeblockError):
-    def __init__(self, info_string):
-        self.info_string = info_string
-
-    def __str__(self):
-        return (
-            "Could not parse the following info string object as"
-            f" valid JSON: '{self.info_string}'"
-        )
